files/requirements.txt
google-generativeai>=0.10.0
streamlit>=1.20
pandas>=1.3
python-dotenv>=1.0
regex>=2023.8.8
requests>=2.28
tqdm>=4.64